Centralize test infrastructure helpers
