+++ Octans.Tests/GlobalUsings.cs
@@ -1,2 +1,3 @@
 global using Octans.Tests;
+global using Octans.Tests.Infrastructure;
 global using Xunit;
