+++ Octans.Tests/Infrastructure/StorageServiceTests.cs
@@ -149,32 +149,4 @@ public class StorageServiceTests
         Assert.Equal("Unknown", result);
     }
 
-    // Helper class that throws exceptions for testing error handling
-    private sealed class ThrowingMockFileSystem : IFileSystem
-    {
-        public IFile File => throw new FakeIoException("Test exception");
-        public IDirectory Directory => throw new FakeIoException("Test exception");
-        public IFileInfoFactory FileInfo => throw new FakeIoException("Test exception");
-        public IFileVersionInfoFactory FileVersionInfo => throw new FakeIoException("Test exception");
-        public IFileStreamFactory FileStream => throw new FakeIoException("Test exception");
-        public IPath Path => throw new FakeIoException("Test exception");
-        public IDirectoryInfoFactory DirectoryInfo => throw new FakeIoException("Test exception");
-        public IDriveInfoFactory DriveInfo => throw new FakeIoException("Test exception");
-        public IFileSystemWatcherFactory FileSystemWatcher => throw new FakeIoException("Test exception");
-    }
-
-    private sealed class FakeIoException : Exception
-    {
-        public FakeIoException(string message) : base(message)
-        {
-        }
-
-        public FakeIoException()
-        {
-        }
-
-        public FakeIoException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-    }
 }
